refactor: keep runtime capabilities behind service boundaries
