+++ docs/internal/baselines/backend-package-file-count.json
@@ -135,7 +135,7 @@
     },
     {
       "packagePath": "pkg/services/factory_sessions/internal/sessionservice",
-      "fileCount": 40,
+      "fileCount": 39,
       "owner": "factory-session-maintainers",
       "removalReason": "Partition session service behavior by lifecycle responsibility."
     },
