+++ pkg/services/automations/internal/service.go
@@ -1,3 +1,10 @@
+// Package internal composes Automations runtime sidecars for script pollers,
+// hosted pollers, filesystem watchers, and cron Work generation. Script
+// command/source polling is owned by internal/services/script_pollers and
+// reached only through this Automations root. Wire supplies explicit
+// submitters, clocks, loggers, cancellation, and configuration. Use
+// StartSchedulerSidecarsForRuntime as the unified runtime entrypoint for poller
+// and cron supervision.
 package internal
 
 import (
@@ -20,6 +27,9 @@ import (
 )
 
 var _ automations.Service = (*Service)(nil)
+
+// WorkRequestSubmitter submits parsed poller or cron work requests into the runtime.
+type WorkRequestSubmitter = automations.WorkRequestSubmitter
 
 // Clock is the automation time source needed for scheduling and supervision.
 type Clock = automations.Clock
