--- pkg/services/automations/internal/types.go
@@ -1,6 +1,0 @@
-package internal
-
-import automations "github.com/portpowered/infinite-you/pkg/services/automations"
-
-// WorkRequestSubmitter submits parsed poller or cron work requests into the runtime.
-type WorkRequestSubmitter = automations.WorkRequestSubmitter
