+++ pkg/services/factory_definitions/internal/service.go
@@ -5,6 +5,7 @@ package internal
 
 import (
 	"context"
+	"fmt"
 
 	factoryroot "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/internal/lifecycle"
@@ -111,4 +112,31 @@ func NewWithAuthoringLayout(
 		distributionService,
 		versionFileSystem,
 	)
+}
+
+type runtimeSnapshotService struct {
+	factoryroot.Service
+	resolve factoryroot.RuntimeSnapshotOperation
+}
+
+// AttachRuntimeSnapshot attaches the owner-composed runtime snapshot
+// operation while preserving the rest of the singular Definitions root.
+func AttachRuntimeSnapshot(
+	service factoryroot.Service,
+	resolve factoryroot.RuntimeSnapshotOperation,
+) (factoryroot.Service, error) {
+	if service == nil {
+		return nil, fmt.Errorf("Factory Definitions service is required")
+	}
+	if resolve == nil {
+		return nil, fmt.Errorf("runtime snapshot operation is required")
+	}
+	return runtimeSnapshotService{Service: service, resolve: resolve}, nil
+}
+
+func (s runtimeSnapshotService) ResolveRuntimeSnapshot(
+	ctx context.Context,
+	request factoryroot.ResolveRuntimeSnapshotRequest,
+) (factoryroot.ResolveRuntimeSnapshotResult, error) {
+	return s.resolve(ctx, request)
 }
