--- pkg/services/factory_definitions/runtime_snapshot_service_test.go
@@ -1,14 +1,0 @@
-package factorydefinitions_test
-
-import (
-	"context"
-
-	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
-)
-
-func (fakeDefinitionsPeer) ResolveRuntimeSnapshot(
-	ctx context.Context,
-	request factorydefinitions.ResolveRuntimeSnapshotRequest,
-) (factorydefinitions.ResolveRuntimeSnapshotResult, error) {
-	return factorydefinitions.UnimplementedService{}.ResolveRuntimeSnapshot(ctx, request)
-}
