--- pkg/services/factory_runtime/runtime_log.go
@@ -1,6 +1,0 @@
-package factory
-
-import "github.com/portpowered/infinite-you/pkg/platform/runtimeartifact"
-
-// RuntimeLogDiagnostics describes immutable runtime log and metrics metadata.
-type RuntimeLogDiagnostics = runtimeartifact.Diagnostics
