--- pkg/services/factory_sessions/directory_inspection_contract.go
@@ -1,5 +1,0 @@
-package factorysessions
-
-// HomeDirectoryResolver resolves the process user's home directory at the
-// external filesystem edge selected by Wire.
-type HomeDirectoryResolver func() (string, error)
