+++ pkg/services/factory_sessions/opened_runtime.go
@@ -8,6 +8,17 @@ import (
 	"github.com/portpowered/infinite-you/pkg/services/workers"
 )
 
+const (
+	ProgressFragmentKind  = workers.ProgressFragmentKind
+	ResponseFragmentKind  = workers.ResponseFragmentKind
+	CompletedFragmentKind = workers.CompletedFragmentKind
+	FailedFragmentKind    = workers.FailedFragmentKind
+)
+
+type ProgressFragment = workers.ProgressFragment
+
+type ProgressPublisher = workers.ProgressPublisher
+
 // SessionRuntimeOpeningRequest contains Factory Session identity,
 // persistence, startup, and hosting values for one runtime.
 type SessionRuntimeOpeningRequest struct {
