+++ pkg/wire/factory_definition_service_provider.go
@@ -1,20 +1,25 @@
 package wire
 
 import (
+	"context"
 	"fmt"
 
 	"github.com/portpowered/infinite-you/pkg/platform/portablefiles"
 	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	factorydefinitionswire "github.com/portpowered/infinite-you/pkg/services/factory_definitions/wire"
+	factoryruntime "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
+	factoryruntimewire "github.com/portpowered/infinite-you/pkg/services/factory_runtime/wire"
 	factorysessions "github.com/portpowered/infinite-you/pkg/services/factory_sessions"
+	factorysessionwire "github.com/portpowered/infinite-you/pkg/services/factory_sessions/wire"
+	"github.com/portpowered/infinite-you/pkg/services/workers"
 )
 
-func provideFactoryDefinitionsRuntimeRouter() factorysessions.DefinitionRuntimeRouter {
-	return factorysessions.NewDefinitionRuntimeRouter()
+func provideFactoryDefinitionsRuntimeRouter() *factorysessions.DefinitionRuntimeRouter {
+	return factorysessionwire.NewDefinitionRuntimeRouter()
 }
 
 func provideFactoryDefinitionsRoot(
-	router factorysessions.DefinitionRuntimeRouter,
+	router *factorysessions.DefinitionRuntimeRouter,
 	validator factorydefinitions.Validator,
 	persistence factorydefinitions.Persistence,
 	loader *factorydefinitionswire.Loader,
@@ -55,4 +60,29 @@ func provideFactoryDefinitionsRoot(
 		portableFileSystem,
 		directoryReplacementStore,
 	)
+}
+
+// provideFactoryRuntimeRootFactory composes the singular process-scoped
+// Runtime root once. Factory Sessions supplies the explicit activation
+// operation when it constructs its opening capability; no late delegate
+// binder is exposed.
+func provideFactoryRuntimeRootFactory(
+	newID factoryruntime.IDGenerator,
+	workflows factoryruntime.JavaScriptWorkflowDefinitions,
+	clock factoryruntime.Clock,
+) factorysessionwire.RuntimeRootFactory {
+	return func(activation factoryruntime.RuntimeActivationOperation) (factorysessionwire.FactoryRuntimeRoot, error) {
+		service, err := factoryruntimewire.NewService(
+			newID,
+			workflows,
+			nil,
+			clock,
+			func(context.Context, workers.WorkstationDispatchRequest) error {
+				return factoryruntime.ErrNotRunning
+			},
+			nil,
+			activation,
+		)
+		return service, err
+	}
 }
